+++ DataMapping/Data/DT_CounterM_Damage.cs
@@ -34,8 +34,8 @@ namespace DataMapping.Data
         {
             DataColumn[] PK_COLUMN = new DataColumn[2];
             TableName = TABLE_NAME;
-            PK_COLUMN[0] = this.Columns.Add(ID_COUNTERM, typeof(int));
-            PK_COLUMN[1] = this.Columns.Add(ID_DAMAGE, typeof(int));
+            PK_COLUMN[0] = this.Columns.Add(ID_COUNTERM, typeof(decimal));
+            PK_COLUMN[1] = this.Columns.Add(ID_DAMAGE, typeof(decimal));
             Columns.Add(COUNTERM_NAMESHORT, typeof(string));
             Columns.Add(TOP_RISK, typeof(string));
             Columns.Add(DAMAGE, typeof(string));
@@ -45,10 +45,10 @@ namespace DataMapping.Data
             Columns.Add(GROUPENAME, typeof(string));
             Columns.Add(STATUS, typeof(bool));
             Columns.Add(VALUE, typeof(decimal));
-            Columns.Add(ID_RISK_TREE, typeof(int));
-            Columns.Add(IDRISK, typeof(int));
-            Columns.Add(ID_GROUPE, typeof(int));
-            Columns.Add(ID_WBS, typeof(int));
+            Columns.Add(ID_RISK_TREE, typeof(decimal));
+            Columns.Add(IDRISK, typeof(decimal));
+            Columns.Add(ID_GROUPE, typeof(decimal));
+            Columns.Add(ID_WBS, typeof(decimal));
             Columns.Add(WBS_NAME, typeof(string));
             Columns.Add(USERNAME, typeof(string));
             PrimaryKey = PK_COLUMN;
