+++ DataMapping/Data/DT_Project.cs
@@ -19,7 +19,7 @@ namespace DataMapping.Data
         {
             DataColumn[] PK_COLUMN = new DataColumn[1];
             this.TableName = TABLE_NAME;
-            PK_COLUMN[0] = this.Columns.Add(ID_PROJECT, typeof(int));
+            PK_COLUMN[0] = this.Columns.Add(ID_PROJECT, typeof(decimal));
             PK_COLUMN[0].AutoIncrement = true; PK_COLUMN[0].AutoIncrementSeed = -100; PK_COLUMN[0].AutoIncrementStep = -1;
             this.Columns.Add(PROJECT_NAME, typeof(string));
             this.PrimaryKey = PK_COLUMN;
